re-use common cli code
